Refactor to use proper header and source files
